requirements.txt
Flask
Flask-JWT-Extended
passlib
python-dotenv
pylint
